Add User#grade and #letter_grade
The grade for a User will be a decimal in the range of 0 to 100.
For now, we're using hard-coded ranges to translate that number into a
letter-grade to display. I assumed that 100 will be the highest chance
that the user will likely become homeless in the near future.

Note that I did NOT allow the grade to be set in the Users controller.
The grade will be recompued using @dummey's machine learning code whenever
the system receives new data about the user.